backend/requirements.txt
Flask
flask-restx
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
pandas
Werkzeug
python-dotenv